Author: Rex Kerr
Fixed a bunch of bugs relating to 64 bit machines--pointer arithmetic using
casts to (int), plus fixed the definition of ub4 in the Jenkins hash
function.  I deleted a number of noprintfs that spit out addresses that we
didn't need any more (I used them for testing only).  MCell3 gives exactly
the same results for the nmj model after the changes, and gives the same
results when compiled in 64 bit mode as in 32 bit mode, even if 4GB of junk
is allocated when MCell3 starts up.

+++ src/sym_table.h
@@ -5,7 +5,7 @@
 
 #ifndef ISAAC64_H
 /* These guys come in for free if we're using Jerkins' random numbers */
-typedef unsigned long int ub4;   /* unsigned 4-byte quantities */
+typedef uint32_t ub4;   /* unsigned 4-byte quantities */
 typedef unsigned char ub1;        /* unsigned 1-byte quantities */
 #endif
